requirements.txt
pandas
dash
dash_bootstrap_components

Flask-login
Flask-Talisman
Flask-sqlalchemy
psycopg2

treelib

pyopenssl

oauthlib
requests

python-dotenv
python-magic
# python-magic-bin
python-docx
openpyxl
XlsxWriter
kaleido
Pillow
